feat(ui): add `uiComponentStore`
So on mobile, when users try to create a group, we don't want to show the "create group" modal, instead we should redirect them to a new route.

Now, this route should only be accessible on mobile. So if users, for example try to be smart (not like anything bad is gonna happen, lol) and manually tye the route's path in the url bar, we redirect them to the dashboard, but this time around with an accompanying `searchParam` that the UI store manages, and we open the modal for them when they land there.