refactor: removed an active state indicator for settings card on mobile
We only need the active state on desktop, since the component is rendered side-by-side the cards, so it is imminent for users to have that visual cue, which we're setting for them either via the `Setting` object or the current URL slug.

The whole point of this refactor is to remove our dependence on `React.useEffect` and passing the `SETTINGS` array as a dependency.