chore(ui): add milestones and rewards base UI components
refactored the dynamic $groupId route to load instantly by removing the loader param which makes it slow because TanStack router tries to preload it first before navigating the navigating this user when the req is successful.

although, this ensures that the data is always available on that route for the use, it affects UX in a way, and it is a trade-off i'm willing to take.