chore(ui): refactor explore filter and milestones UI (#78)
* Explore filters (#73)

* feat(ui): done with the explore page filters

* fix: resolved the issues highlighted

* fix(ci): removed unneccessary dependencies

* fix(ui): ensure startDate is included in query params and improve UI feedback

* chore(ui): add milestones and rewards base UI components

refactored the dynamic $groupId route to load instantly by removing the loader param which makes it slow because TanStack router tries to preload it first before navigating the navigating this user when the req is successful.

although, this ensures that the data is always available on that route for the use, it affects UX in a way, and it is a trade-off i'm willing to take.

* chore: make loading prop in GroupCard optional

---------

Co-authored-by: Jettechnologies <80001133+jettechnologies@users.noreply.github.com>